Changed "DeltaHyperbolicityToolkit" to "Graph" in a separate solution, as it only contains graph-related code, completely unrelated to hyperbolicity stuff... (currently it still has a few things that need to move). The DeltaHyperbolicity solution has the program that uses this Graph code to run Algorithms (the 3rd solution) to find delta-hyperbolicity of graphs.
*Also small bug fix - can now handle the case where a full path is given for a shared library.

+++ Graph/src/Graph/GraphAlgorithms.cpp
@@ -388,7 +388,7 @@ namespace dhtoolkit
 			delta = cycleDelta(cycleSize);
 			return false;
 		}
-		else if (stDist < chain1.size() + chain2.size() + 2)
+		else if (static_cast<size_t>(stDist) < chain1.size() + chain2.size() + 2)
 		{
 			//calculate cycle size
 			size_t cycleSize = chain1.size() + chain2.size() + 2 + stDist;
